Run SQL script from the command line and export as TSV
From within MySQL, export to CSV (SELECT ... INTO OUTFILE ...) requires
an absolute path and write privileges on the file system.

I favor this solution for now, but might have to reconsider if field
escaping or character set encoding introduce extra complexity.

In this case, I would export to a temporary file with a unique name
in /tmp, then move the file to the 'output' folder of the project.