Add column 'institutions' to export
The name of each institution is quoted in double quote characters '"'.

I noted that commas are present in the names of institutions,
with no space between the comma and the following character.

I expect this defect to be fixed upstream, or to apply a fix
to the data before running this export query.