Add column 'institution_types' to export
Like in the 'institutions' column, the values in 'institution_types'
are quoted with double quotes.

For these two columns, only distinct values are listed, without keeping
a reference to which assessment report the author contributed to with each
institution. The relationship between an institution and its type is lost
as well: for each author, institution types are listed independently and
possibly in a different order from institutions.