Use greater than or equal with total number of contributions
In previous version, authors selected when their total number
of contributions matched exactly the value of a category. In
this corrected version, authors are selected when their total
number of contributions is greater than or equal to the value
of a category. As a result, authors with multiple contributions
will now be listed in more than one category.