Add answer to question 2 for working groups
Note:
I expected 3 authors in 5+, but found only 2 in the answer.

The 3 authors are:
  * Gerald A. Meehl (5 AR in WG I)
  * Ronald J. Stouffer (4 AR in WG I, 1 AR in WG III)
  * Kevin E. Trenberth (5 AR in WG I)

Which color should Ronald J. Stouffer have in the bar chart,
in the bar for authors with 5+ AR participations?

To avoid this issue, the dimension of the bar could be based
on the total number of participations, not the total number
of authors. But this makes the chart more theoretical and harder
to comprehend.

Thus, I take note of and accept the discrepancy: the dimensions of
the bars will vary depending on the colors applied, which is due
to differences based on selected grouping.