Add extra answer to explicit results of different answers
The total number of authors with 5 participations is smaller,
when counting by institution, then the same count in general.

This is because the count by institution only counts participations
of this author for this institution (a relevant choice), which is
often the same but may be less than the total number of participations
of the author in general.