Add SQL query to answer question 15
This is hierarchic data in nature, built with joins
with parent groups repeated in each result row.

This data is intended to be represented in treemaps,
with symbol columns only rendered with colors,
while other columns will display text.

The AR column is intended for aggregation and filtering:
a range of AR may be selected, and the corresponding
aggregate count of participations will be computed.