Add missing multiplication factor for bridge groups: x2, x3
In previous ratio, only one of the groups part of a bridge
was counted, instead of all groups (up to the minimum of the
participations of the groups part of the bridge).

Due to rounding errors, some ratios appear off,
e.g. 0.99999 instead of 1. To reduce these rounding
errors, the division by the total number of participations
should be delayed until the last step.